+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- update: Firebase, Blueprint, and other dependencies.
+- update: Firebase, Blueprint, Webpack, and other dependencies.
 
 
 ## [4.11.0] - 2018-06-20