+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 - update: Firebase, Blueprint, Webpack, and other dependencies.
 - feature: <SimpleEditor /> component added.
+- feature: Image captions can have marks (bold, italic, superscript, subscript, code).
+- fix: NodeView losing focus on typing while in Collaborative mode fixed.
 
 ## [4.11.0] - 2018-06-20
 - feature: Add importHtml function to <Editor />.